FEAT: big improvement on the lib with type changes on functions and data structures to be cleaner

+++ libs/my_c/src/hash/my_create_table.c
@@ -19,7 +19,7 @@
  * @return <b>hashtable_t*</b> The created hashtable
  * @author Nicolas TORO
  */
-hashtable_t *my_create_hashtable(int (*hash)(char *, int), int len)
+hashtable_t *my_create_hashtable(ssize_t (*hash)(char *, size_t), size_t len)
 {
     hashtable_t *hashtable;
 
@@ -31,7 +31,7 @@ hashtable_t *my_create_hashtable(int (*hash)(char *, int), int len)
     hashtable->hash = hash;
     hashtable->len = len;
     hashtable->table = my_calloc(len, sizeof(hashtable_entry_t *));
-    for (int i = 0; i < len; i++)
+    for (size_t i = 0; i < len; i++)
         hashtable->table[i] = NULL;
     return hashtable;
 }
